% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSA.R
\name{geneSA}
\alias{geneSA}
\title{Tool for Identification of genes significantly associated with prognostic value (survival rates of patients)}
\usage{
geneSA(genename = NULL, event = NULL)
}
\arguments{
\item{genename, event}{}
}
\description{
a log-rank test in univariate Cox regression analysis with a proportional hazards model is performed to examine the association between the expression of each gene and the survival rates of patients. Genes with Q-value <= 0.05 (Benjamini-Hocberg procedure) are preserved. For further information on requirements as well as how to implement this tool, please visit my Github repository: https://github.com/huynguyen250896/geneSA
}
\examples{
geneSA(genename = vt, event = df)

}
