% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSA.R
\name{geneSA}
\alias{geneSA}
\title{geneSA: Identification of genes significantly associated with patient outcome.}
\usage{
geneSA(data, time, status, Pcut, Qcut)
}
\arguments{
\item{data}{data frame or matrix. It represents its rows are samples and its columns are genomic features.
Note that samples in rows of \code{data} are also included in your clinical data and in exactly the same order.}
  
\item{time}{time numeric or integer column vector. It is overall survival time of all samples extracted from 
your clinical data. Note that samples in rows of clinical data are included in \code{data} and in exactly the 
same order before extracting it.}

\item{status}{binary column vector. It is overall survival status of all samples extracted from your clinical 
data (usually coded as 1 = death, 0 = alive). Note that samples in rows of clinical data are included in \code{data} 
and in exactly the same order before extracting it.}

\item{Pcut}{numeric. A user-defined P-value threshold to define statistical significance level. Default value 
is P-value <= 0.05.}

\item{Qcut}{numeric. A user-defined Q-value threshold to define statistical significance level. Default value 
is Q-value <= 0.05.}

\item{univariate}{boolean. Whether \code{geneSA} runs an univariate or a multivariate survival analysis. Default value is univariate = T}
}

\description{
a log-rank test in univariate Cox regression analysis with a proportional hazards model is performed to examine 
the association between each gene and patient outcome. Genes with Q-value <= 0.05 (Benjamini-Hocberg procedure) 
are preserved. More detailed information on requirements of as well as how to implement this tool, please visit my 
Github repository: https://github.com/huynguyen250896/geneSA
}

\examples{
geneSA(data = exp1, time = clinical_exp$OS_MONTHS, status = clinical_exp$status, Pcut = 0.05, Qcut = 0.05, univariate = TRUE)
}

\references{
Quang-Huy Nguyen, Duc-Hau Le. (2020). 
Improving existing analysis pipeline to identify and analyze cancer driver genes using multi-omics data.
Scientific Reports, 10(1):20521.
}  
  
\author{
Quang-Huy Nguyen
}
